Modify _read function in MemDuplex.
Push into readable buffer the first item in the
memDuplex buffers array rather than keeping track of current.

The behavior of this.push on the readable stream was not
consistent which was resulting in this.cur to be over incremented
and for some chunks to be pushed into the readable stream buffer
multiple times.