Add custom stream error handling
Provide an optional error handler function as the last parameter to
readySetStream(), called when an error happens on the way to stream
output with the error object as parameter. The idea is to allow more
specific error handling than the default behavior.

The default error handler if none is provided calls destroy() on the
connection object in the response.